adds showModal() and closeModal() to App.js, passes as props to BookFormModal.js
